+++ STACpopulator/populators.py
@@ -11,7 +11,7 @@ import os
 import re
 from abc import ABC, abstractmethod
 from datetime import datetime
-from typing import Any, Callable, Dict, Iterable, List, MutableMapping, Optional, Type, Union, get_args
+from typing import Any, Callable, Dict, Iterable, List, MutableMapping, Optional, Union, get_args
 
 import pystac
 import requests
@@ -26,7 +26,6 @@ from STACpopulator.api_requests import (
 from STACpopulator.collection_update import UpdateModesOptional, update_collection
 from STACpopulator.exceptions import FunctionLoadError
 from STACpopulator.input import ErrorLoader, GenericLoader, THREDDSLoader
-from STACpopulator.models import AnyGeometry
 from STACpopulator.stac_utils import load_config
 
 LOGGER = logging.getLogger(__name__)
@@ -164,12 +163,6 @@ class STACpopulatorBase(ABC):
     def collection_id(self) -> str:
         """Return the populator's collection id."""
         return self._collection_info["id"]
-
-    @property
-    @abstractmethod
-    def item_geometry_model(self) -> Type[AnyGeometry]:
-        """Return a geometry model class that represents the geometry used in this populator."""
-        raise NotImplementedError
 
     @abstractmethod
     def create_stac_item(self, item_name: str, item_data: dict[str, Any]) -> dict[str, Any]:
@@ -420,6 +413,23 @@ class THREDDSPopulator(STACpopulatorBase):
                 "By default, uses the adjacent configuration to the implementation class."
             ),
         )
+        parser.add_argument(
+            "--fallback-crs",
+            help=(
+                "Coordinate reference system (CRS) to use if not specified by the NCML metadata. "
+                "This can either be specified as any string input to the pyproj.CRS class. "
+                "Use this if the NCML does not contain the geospatial_bounds_crs CF metadata field "
+                "to indicate the CRS used for this data."
+            ),
+        )
+        parser.add_argument(
+            "--force-crs",
+            help=(
+                "Coordinate reference system (CRS) to use. This will override the CRS specified in the metadata. "
+                "This can either be specified as any string input to the pyproj.CRS class. "
+                "Use this to force the use of a different CRS if the one specified in the NCML metadata is incorrect."
+            ),
+        )
 
     @classmethod
     def run(cls, ns: argparse.Namespace, session: requests.Session) -> int:
@@ -427,7 +437,7 @@ class THREDDSPopulator(STACpopulatorBase):
         LOGGER.info(f"Arguments to call: {vars(ns)}")
 
         if ns.mode == "full":
-            data_loader = THREDDSLoader(ns.href, session=session)
+            data_loader = THREDDSLoader(ns.href, fallback_crs=ns.fallback_crs, force_crs=ns.force_crs, session=session)
         else:
             # To be implemented
             data_loader = ErrorLoader()